requirements.txt
asgiref==3.3.2
Django==3.2
django-cors-headers==3.6.0
django-simple-history==2.12.0
djangorestframework==3.12.2
drf-history==0.0.9
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
